m12_10_additional/pyproject.toml
[tool.poetry]
name = "m12-10-additional"
version = "0.1.0"
description = ""
authors = ["Yuriy Kuchma <krabatua@gmail.com>"]
readme = "README.md"
packages = [{include = "m12_10_additional"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
pymongo = "^4.4.1"
beautifulsoup4 = "^4.12.2"
redis = "^4.6.0"
celery = "^5.3.1"
requests = "^2.31.0"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
